move useAuthed into user
